test: Calculator의 수익율 계산 로직 테스트 함수 작성 및 테스트가 통과하도록 수정

+++ src/main/kotlin/lotto/data/Stats.kt
@@ -2,7 +2,7 @@ package lotto.data
 
 data class Stats(
     val info: WinningInfo,
-    val profitRate: Float,
+    val profitRate: Double,
 ) {
     override fun toString(): String {
         return """
